The OffersList component has been rewritten to improve clarity
